+++ main.py
@@ -13,9 +13,9 @@ Licensed under the MIT License (see LICENSE for details)
 Usage:
 refer to README.md file
 """
-from modules.parser import *
-from modules.bounding_boxes import *
-from modules.image_level import *
+from oidv4_toolkit.parser import *
+from oidv4_toolkit.bounding_boxes import *
+from oidv4_toolkit.image_level import *
 
 
 ROOT_DIR = ''
