+++ oidv4_toolkit/csv_downloader.py
@@ -4,7 +4,7 @@ import time
 import urllib.request
 import pandas as pd
 
-from modules.utils import bcolors as bc
+from oidv4_toolkit.utils import bcolors as bc
 
 OID_URL = 'https://storage.googleapis.com/openimages/2018_04/'
